refactor(store): derive ToolCallRow::ok from state instead of storing both
Two struct fields that must agree are a disagreement waiting to be
introduced by the next person who sets one of them. The table keeps both
columns (`ok` for every pre-v18 reader, `state` for the lifecycle); in
memory there is now one field and one answer.

Moves recent_tool_calls' test to the module that owns it, keeping
tests.rs under its file-size ceiling.
